Refactor: Separate schemas, models, and services
This refactoring separates the application's concerns into three distinct layers:

1.  **schemas**: A new `src/schemas` directory now holds the Pydantic models (DTOs) that define data structures, such as `FileChange` and `SearchRequest`.

2.  **models**: The `src/models` directory is now the home for stateful classes with behavior, such as `GitManager`, `VectorStore`, and `ObsidianProcessor`. These were previously in `src/services`.

3.  **services**: The `src/services` directory now contains only the use-case-specific orchestrators, like `SyncCoordinator`, which coordinate multiple models.

This change clarifies the architecture and improves the separation of concerns. All import paths across the application and test suite have been updated to reflect this new structure, and `__init__.py` files have been configured to provide simplified imports.
