+++ src/models/obsidian_processor.py
