+++ src/schemas/__init__.py
@@ -1,0 +1,5 @@
+"""Schemas for the application."""
+
+from .app_schemas import FileChange, SearchRequest, SearchResult
+
+__all__ = ["FileChange", "SearchRequest", "SearchResult"]
